+++ cdk/lib/s2s-stack.ts
@@ -110,7 +110,15 @@ export class S2SAppStack extends cdk.Stack {
 
     const dynamoDbReadPolicy = new iam.PolicyStatement({
       effect: iam.Effect.ALLOW,
-      actions: ["dynamodb:GetItem", "dynamodb:BatchGetItem", "dynamodb:Scan", "dynamodb:Query", "dynamodb:PutItem", "dynamodb:UpdateItem", "dynamodb:DeleteItem"],
+      actions: [
+        "dynamodb:GetItem",
+        "dynamodb:BatchGetItem",
+        "dynamodb:Scan",
+        "dynamodb:Query",
+        "dynamodb:PutItem",
+        "dynamodb:UpdateItem",
+        "dynamodb:DeleteItem",
+      ],
       resources: [
         `arn:aws:dynamodb:${this.region}:${this.account}:table/${dynamoDbTableName}`,
         `arn:aws:dynamodb:${this.region}:${this.account}:table/${dynamoDbTableName}/index/*`,
@@ -204,43 +212,6 @@ export class S2SAppStack extends cdk.Stack {
         unhealthyThresholdCount: 2,
       },
     });
-  }
-
-  // Temporary workaround to force credential refresh every 5 hours by stopping tasks
-  // (until official Python SDK supports automatic credential fetching from ECS task role)
-  temp_addTasksRotateLambda(
-    ecsService: ecs.FargateService,
-    ecsCluster: ecs.Cluster
-  ) {
-    const taskRotater = new NodejsFunction(this, "ECSTaskRotateLambda", {
-      runtime: lambda.Runtime.NODEJS_LATEST,
-      entry: path.join(__dirname, "../lambda/ecsTaskRotater/index.ts"),
-      handler: "handler",
-      timeout: Duration.minutes(10),
-      environment: {
-        ECS_CLUSTER_NAME: ecsCluster.clusterName,
-        ECS_SERVICE_NAME: ecsService.serviceName,
-      },
-    });
-    taskRotater.addToRolePolicy(
-      new iam.PolicyStatement({
-        actions: ["ecs:ListTasks", "ecs:StopTask"],
-        resources: ["*"],
-        conditions: {
-          ArnEquals: {
-            "ecs:cluster": ecsCluster.clusterArn,
-          },
-        },
-      })
-    );
-    // Create EventBridge rule to trigger the Lambda every 5 hours: before credentials expire (6 hours)
-    const rule = new events.Rule(this, "ScheduleTaskRotation", {
-      schedule: events.Schedule.rate(Duration.hours(5)),
-      description: "Triggers ECS task rotation every 5 hours",
-    });
-
-    // Add the Lambda function as a target for the rule
-    rule.addTarget(new targets.LambdaFunction(taskRotater));
   }
 
   createFrontendConfig() {
@@ -358,8 +329,6 @@ export class S2SAppStack extends cdk.Stack {
     );
     const wsService = this.createContainerDefinition(ecsCluster, wsTaskDef);
     this.registerLoadBalancerTarget(wsService);
-    // This is a temporary workaround to the lack of credential refresh in the experimental Python SDK
-    this.temp_addTasksRotateLambda(wsService, ecsCluster);
 
     this.createFrontendConfig();