+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adapt to an official version of the async clients for the [AWS SDK for Python](https://github.com/awslabs/aws-sdk-python/tree/develop) when the experimental one becomes official.
 - [developer experience] Easier custom tool extension (through some framework)
 
+## [0.1.5] - 2024-10-30
+
+- Upgraded from aws_sdk_bedrock_runtime 0.0.2 to 0.1.1
+- The new SDK version improved resource utilization which was a key issue for many users (see: https://github.com/awslabs/aws-sdk-python/issues/11)! You should see decreased CPU load with a bidirectional stream open. 🎉
+- Removed old workaround for lack of ContainerCredentialsResolver which involved a lambda restarting tasks every 5 hours. 
+- The new code dynamically picks between ContainerCredentialsResolver and EnvironmentCredentialsResolver depending on whether you are running the application locally or with backend deployed on ECS.
+- Fixed bug where running locally would lead to a no region found error when initializing boto3 clients for the knowledge base search or user lookup tool
+
+
 ## [0.1.4] - 2024-10-23 (hotfix)
 
 - Pinned AWS Smithy SDK dependencies to specific versions to prevent breaking changes from experimental 0.1.0 packages (`smithy-core==0.0.2`, `aws-sdk-signers==0.0.3`, `aws_sdk_bedrock_runtime==0.0.2`, `smithy-aws-core==0.0.3`, `smithy-http==0.0.1`, `smithy-json==0.0.1`, `smithy-aws-event-stream==0.0.1`)